requirements.txt
fastapi
python-jose
pydantic
httpx
orjson
redis
uvicorn
fastapi-utils
typing-inspect
fastapi-lifespan-manager